finished front end listener for edit tasks

+++ routes/api-notes.js
@@ -2,7 +2,7 @@ const express = require("express");
 const apiNoteRouter = express.Router();
 const { requireAuth } = require("../auth");
 const { Note, Task } = require("../db/models");
-const { asyncHandler, deleteItem } = require("./utils");
+const { asyncHandler, deleteItem, findCurrentUser } = require("./utils");
 
 
 apiNoteRouter.delete(
@@ -25,6 +25,20 @@ apiNoteRouter.delete(
 
 apiNoteRouter.post("/", requireAuth, asyncHandler(async (req, res) => {
   const { content, userId, taskId } = req.body;
+
+  try{
+    await Note.create({ content, userId, taskId });
+  } catch (err) {
+    console.err("messed up on backend note create", err);
+  }
+  const notes = await Note.findAll({ where: { taskId } });
+  res.json(notes)
+}))
+
+
+apiNoteRouter.patch("/:id", requireAuth, asyncHandler(async (req, res) => {
+  const { content, taskId } = req.body;
+  const userId =
 
   try{
     await Note.create({ content, userId, taskId });
